Refactor FirmsController & TradingNamesController into abstract class.
We want to edit and update both, but we only want to index firms and we
only want to create trading names. So it makes sense to pull this back.
