+++ spec/controllers/dashboard/firms_controller_spec.rb
@@ -9,7 +9,17 @@ module Dashboard
     let(:user) { FactoryGirl.create :user, principal: firm.principal }
     before { sign_in(user) }
 
-    describe 'GET index' do
+    def extract_firm_params(firm, params = {})
+      firm_params = firm.attributes
+      firm_params.symbolize_keys!
+      firm_params[:initial_advice_fee_structure_ids] = firm.initial_advice_fee_structure_ids
+      firm_params[:ongoing_advice_fee_structure_ids] = firm.ongoing_advice_fee_structure_ids
+      firm_params[:allowed_payment_method_ids] = firm.allowed_payment_method_ids
+      firm_params[:investment_size_ids] = firm.investment_size_ids
+      firm_params.merge(params)
+    end
+
+    describe 'GET #index' do
       context 'when all trading names are registered' do
         it 'assigns all trading names' do
           get :index
@@ -27,6 +37,49 @@ module Dashboard
         it 'assigns only registered trading names' do
           get :index
           expect(assigns(:trading_names).count).to eq 2
+        end
+      end
+    end
+
+    describe 'GET #edit' do
+      let(:trading_name) { FactoryGirl.create(:firm, parent_id: firm.id) }
+      before { get :edit, id: firm.id }
+
+      it 'assigns the firm' do
+        expect(assigns(:firm)).to eq firm
+      end
+
+      it 'renders the edit page' do
+        expect(response).to render_template 'dashboard/firms/edit'
+      end
+    end
+
+    describe 'PATCH #update' do
+      let(:firm) { FactoryGirl.create(:firm) }
+
+      context 'when passed valid details' do
+        let(:firm_params) { extract_firm_params(firm, email_address: 'valid@example.com') }
+        before { patch :update, id: firm.id, firm: firm_params }
+
+        it 'updates the firm' do
+          expect(firm.reload.email_address).to eq firm_params[:email_address]
+        end
+
+        it 'renders the edit page' do
+          expect(response).to render_template 'dashboard/firms/edit'
+        end
+      end
+
+      context 'when passed invalid details' do
+        let(:firm_params) { extract_firm_params(firm, email_address: 'not_valid') }
+        before { patch :update, id: firm.id, firm: firm_params }
+
+        it 'does not update the firm' do
+          expect(firm.reload.email_address).not_to eq firm_params[:email_address]
+        end
+
+        it 'renders the edit page' do
+          expect(response).to render_template 'dashboard/firms/edit'
         end
       end
     end
